Checking in my attempts at parsing Pro Tools cloud projects.
PT cloud projects seem to follow these conventions:
1. They're stored in a folder with some long name (hash?) ending in ".ptproject", i.e "1onItLzyb9gKyQAUYP0Xkb#.ptproject"
2. Inside that folder there's a ".ptxl" file and a "Revision History" folder (along with some other PT files).
3. The ".ptxl" file contains the relative path to the latest revision (i.e "Revision History/Manual Saves/20200715.023322.ptpx")

I Think the ".ptpx" file is actually the one that contains the session info, but I'm not sure how to debug it. I got as far as seeing an 'xor_type' of 0x06in the unxor() function, but after that point I wasn't sure how to decrypt the file.